feat(client): convert ProfilePage to floating modal panel
Replace the standalone /profile route with a globally-rendered modal
overlay. The UserCapsule dropdown now opens the ProfileModal directly
instead of navigating. Visiting /profile redirects to /lobby and opens
the modal for backward compatibility.
